slipt unimp_trainer to three file
